+++ src/resources/assets/js/components/annotationShapes/Circle.vue
