+++ src/resources/assets/js/components/annotationShapes/Ellipse.vue
