+++ src/resources/assets/js/components/annotationShapes/Line.vue
