+++ src/resources/assets/js/components/annotationShapes/Polygon.vue
