+++ src/resources/assets/js/components/annotationShapes/Rectangle.vue
