+++ src/resources/assets/js/components/annotationShapes/mixins/Shape.vue
